finished writing files, just need to fix line approximation and order validation

+++ src/main/java/model/FlightPath.java
@@ -4,8 +4,8 @@ public record FlightPath(
         String orderNo,
         double fromLongitude,
         double fromLatitude,
-        double angle,
+        Double angle,
         double toLongitude,
         double toLatitude,
-        int ticksSinceStartOfCalculation
+        long ticksSinceStartOfCalculation
 ) {}
